+++ src/main/java/de/maxhenkel/persistentplayers/entities/PersistentPlayerEntity.java
@@ -25,7 +25,9 @@ import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.world.World;
 import net.minecraft.world.server.ServerWorld;
+import net.minecraftforge.fml.common.ObfuscationReflectionHelper;
 
+import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.Optional;
 import java.util.UUID;
@@ -69,6 +71,7 @@ public class PersistentPlayerEntity extends MobEntity {
         persistentPlayer.setFire(player.func_223314_ad());
         player.getActivePotionEffects().forEach(persistentPlayer::addPotionEffect);
         persistentPlayer.setInvulnerable(player.isCreative());
+        persistentPlayer.setPlayerModel(getModel(player));
         return persistentPlayer;
     }
 
@@ -84,6 +87,17 @@ public class PersistentPlayerEntity extends MobEntity {
         player.prevRotationPitch = prevRotationPitch;
         player.rotationYawHead = rotationYawHead;
         player.prevRotationYawHead = prevRotationYawHead;
+    }
+
+    public static byte getModel(PlayerEntity player) {
+        try {
+            Field flag = ObfuscationReflectionHelper.findField(PlayerEntity.class, "field_184827_bp");
+            DataParameter<Byte> dataParameter = (DataParameter<Byte>) flag.get(null);
+            return player.getDataManager().get(dataParameter);
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+            return 0;
+        }
     }
 
     @Override