+++ src/main/java/de/maxhenkel/persistentplayers/Main.java
@@ -11,7 +11,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class Main {
 
     public static final String MODID = "persistent_players";
-    public static final String VERSION = "1.12.2-1.0.0";
+    public static final String VERSION = "1.12.2-1.0.1";
     public static final String MC_VERSION = "[1.12.2]";
     public static final String UPDATE_JSON = "https://maxhenkel.de/update/persistent_players.json";