+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed model not applied to offline player
+- Updated to 1.16.4